+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
-- Generate typings for class constructors
+- Generate typings for class constructors.
+- Add `@ts-ignore` comment if method contains `@suppress {checkTypes}` in jsdoc.
 
 ## [1.2.2] - 2018-03-19
 - Fix bad file path handling which broke Windows support.